Started work on the Playoffs
LaplaceSim.py: Added a variable to track which team lost, makes it easier to know which team to eliminate
Standings.py: Created a function to make the seeding for the playoffs
mlbMain.py: Creates and runs the actual games in the playoffs.  So far, it only runs WildCard games.  Needs refactoring.