Finished the Playoffs simulation.
Next step(s): create a main function that will ask for user input.  Simulation can either be 1 season or simming the season multiple times.  In the case of simming the season multiple times, information will need to be stored and than printed at the end of all the simulations.  Would also like to add a home field advantge to the simualtion of the games